FIX: Fix definition of MappingIntervalDict
