+++ python/python/databento_dbn/__init__.py
@@ -1,11 +1,30 @@
 import datetime as dt
-from typing import Protocol, Sequence
+from typing import Protocol, Sequence, TypedDict
 
 # Import native module
 from ._lib import *  # noqa: F403
 
 
 class MappingInterval(Protocol):
+    """
+    Represents a symbol mapping over a start and end date range interval.
+
+    Parameters
+    ----------
+    start_date : dt.date
+        The start of the mapping period.
+    end_date : dt.date
+        The end of the mapping period.
+    symbol : str
+        The symbol value.
+
+    """
+
+    start_date: dt.date
+    end_date: dt.date
+    symbol: str
+
+class MappingIntervalDict(TypedDict):
     """
     Represents a symbol mapping over a start and end date range interval.
