+++ python/python/databento_dbn/_lib.pyi
@@ -5,8 +5,9 @@ import datetime as dt
 from collections.abc import Iterable
 from collections.abc import Sequence
 from enum import Enum
-from typing import BinaryIO, ClassVar, SupportsBytes, TextIO, TypedDict, Union
+from typing import BinaryIO, ClassVar, SupportsBytes, TextIO, Union
 
+from databento_dbn import MappingIntervalDict
 from databento_dbn import SymbolMapping
 
 
@@ -230,24 +231,6 @@ class VersionUpgradePolicy(Enum):
     def from_str(cls, value: str) -> SType: ...
     @classmethod
     def variants(cls) -> Iterable[SType]: ...
-
-class MappingIntervalDict(TypedDict):
-    """
-    Represents a symbol mapping over a start and end date range interval.
-
-    Parameters
-    ----------
-    start_date : dt.date
-        The start of the mapping period.
-    end_date : dt.date
-        The end of the mapping period.
-    symbol : str
-        The symbol value.
-    """
-
-    start_date: dt.date
-    end_date: dt.date
-    symbol: str
 
 class Metadata(SupportsBytes):
     """
